Revert "@LACHAIN-0 refactor ConsensusManager flow to dispose era correctly"
This reverts commit 7bba518a2c4fb5b10e9180d3f41fb2a4068add97.

+++ src/Lachain.Console/Application.cs
@@ -111,7 +111,7 @@ namespace Lachain.Console
                     .Where(key => !key.Equals(wallet.EcdsaKeyPair.PublicKey))
             );
             Logger.LogInformation("Block synchronization finished, starting consensus...");
-            consensusManager.Start(blockManager.GetHeight() + 1);
+            consensusManager.Start((long) blockManager.GetHeight() + 1);
             validatorStatusManager.Start(false);
 
             System.Console.CancelKeyPress += (sender, e) =>
